# docs/openapi.json
{"openapi":"3.0.0","paths":{"/api/v1/health":{"get":{"operationId":"HealthController_getHealth","parameters":[],"responses":{"200":{"description":"The Health Check is successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"ok"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"error"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"},"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}}},"tags":["Health"]}},"/api/v1/companies/{crn}/settlement-eligibility":{"get":{"operationId":"CompaniesController_getEligibility","parameters":[],"responses":{"200":{"description":""}},"tags":["Companies"]}},"/api/v1/settlement-requests":{"post":{"operationId":"SettlementRequestsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSettlementRequestDto"}}}},"responses":{"201":{"description":""}},"tags":["SettlementRequests"]},"get":{"operationId":"SettlementRequestsController_list","parameters":[],"responses":{"200":{"description":""}},"tags":["SettlementRequests"]}},"/api/v1/settlement-requests/{id}":{"get":{"operationId":"SettlementRequestsController_detail","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["SettlementRequests"]}},"/api/v1/backoffice/settlement-requests":{"get":{"operationId":"BackofficeController_list","parameters":[{"name":"status","required":false,"in":"query","schema":{"type":"string","enum":["SUBMITTED","UNDER_REVIEW","REJECTED","AWAITING_PAYMENT","PAYMENT_PROCESSING","PAID","PARTIALLY_SETTLED","SETTLED"]}},{"name":"cursor","required":false,"in":"query","description":"Opaque cursor returned by the previous page.","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","schema":{"minimum":1,"maximum":100,"default":25,"allOf":[{"$ref":"#/components/schemas/Object"}]}},{"name":"crn","required":false,"in":"query","schema":{"example":"CRN-DEMO-001","type":"string"}},{"name":"from","required":false,"in":"query","schema":{"format":"date-time","type":"string"}},{"name":"to","required":false,"in":"query","schema":{"format":"date-time","type":"string"}}],"responses":{"200":{"description":""}},"tags":["Backoffice"]}},"/api/v1/backoffice/settlement-requests/{id}":{"get":{"operationId":"BackofficeController_detail","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Backoffice"]}},"/api/v1/backoffice/settlement-requests/{id}/fees":{"patch":{"operationId":"BackofficeController_setFees","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SetFeesDto"}}}},"responses":{"200":{"description":""}},"tags":["Backoffice"]}},"/api/v1/backoffice/settlement-requests/{id}/approve":{"post":{"operationId":"BackofficeController_approve","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Backoffice"]}},"/api/v1/backoffice/settlement-requests/{id}/reject":{"post":{"operationId":"BackofficeController_reject","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RejectRequestDto"}}}},"responses":{"201":{"description":""}},"tags":["Backoffice"]}},"/api/v1/settlement-requests/{id}/payment-summary":{"get":{"operationId":"PaymentsController_summary","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Payments"]}},"/api/v1/settlement-requests/{id}/payment":{"post":{"operationId":"PaymentsController_pay","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PayDto"}}}},"responses":{"201":{"description":""}},"tags":["Payments"]}},"/api/v1/notifications":{"get":{"operationId":"NotificationsController_list","parameters":[],"responses":{"200":{"description":""}},"tags":["Notifications"]}},"/api/v1/notifications/unread-count":{"get":{"operationId":"NotificationsController_unread","parameters":[],"responses":{"200":{"description":""}},"tags":["Notifications"]}},"/api/v1/notifications/{id}/read":{"patch":{"operationId":"NotificationsController_read","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Notifications"]}},"/api/v1/backoffice/meetings/meeting-attachment":{"post":{"operationId":"DocumentsController_uploadAttachment","parameters":[],"responses":{"201":{"description":""}},"tags":["Documents"]}},"/api/v1/backoffice/meetings/{id}":{"get":{"operationId":"DocumentsController_download","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Documents"]}},"/api/v1/backoffice/meetings/{id}/settlement-document":{"post":{"operationId":"DocumentsController_upload","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Documents"]}}},"info":{"title":"MCDR Settlement Service","description":"Secure settlement workflow API","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateMeetingDto":{"type":"object","properties":{"meetingAt":{"format":"date-time","type":"string","example":"2025-01-15T10:00:00.000Z"},"capital":{"type":"number","example":1000000,"minimum":0.01},"attachmentDocumentId":{"type":"string","format":"uuid"}},"required":["meetingAt","capital","attachmentDocumentId"]},"CreateSettlementRequestDto":{"type":"object","properties":{"crn":{"type":"string","example":"CRN-DEMO-001"},"meetings":{"minItems":1,"maxItems":20,"type":"array","items":{"$ref":"#/components/schemas/CreateMeetingDto"}}},"required":["crn","meetings"]},"Object":{"type":"object","properties":{}},"MeetingFeeDto":{"type":"object","properties":{"meetingId":{"type":"string","format":"uuid"},"amount":{"type":"number","example":1500,"minimum":0.01}},"required":["meetingId","amount"]},"SetFeesDto":{"type":"object","properties":{"fees":{"minItems":1,"type":"array","items":{"$ref":"#/components/schemas/MeetingFeeDto"}}},"required":["fees"]},"RejectRequestDto":{"type":"object","properties":{"reason":{"type":"string","example":"Attachment is not legible."}},"required":["reason"]},"PayDto":{"type":"object","properties":{"idempotencyKey":{"type":"string","minLength":8,"maxLength":255,"example":"payment-attempt-2025-0001"}},"required":["idempotencyKey"]}}}}